Update ExtUtils-MakeMaker to CPAN version 7.28
  [DELTA]

7.28  Tue May 30 22:01:08 BST 2017

    No changes since v7.27_02

7.27_02 Tue May 30 09:27:56 BST 2017

    Bug fixes:
    - Prune auto/share from search paths RT#121918

7.27_01 Sun May 28 11:35:46 BST 2017

    Bug fixes:
    - Fix regression with metadata RT#121913

7.26  Sat May 27 21:01:47 BST 2017

    No changes since 7.25_06

7.25_06 Tue May 23 20:18:01 BST 2017

    Bug fixes:
    - Fix regression with XS tests on MSWin32 with MS toolchain

7.25_05 Mon May 15 10:18:01 BST 2017

    Bug fixes:
    - Make MakeMaker pass compilation tests on AIX again
    - Test, fix test dep on SKIPped linktype

7.25_04 Fri May 12 12:24:09 BST 2017

    Enhancements:
    - Add os_unsupported() function

7.25_03 Thu May 11 17:51:23 BST 2017

    Bug fixes:
    - processPL now depends on 'pure_all' instead of 'all'

7.25_02 Thu May 11 11:54:42 BST 2017

    Bug fixes:
    - Only add staticlibs that are installed under auto/
    - Correct the order of tests of chmod()

    Doc fixes:
    - Fixed typo in MakeMaker.pm

7.25_01 Fri Feb  3 13:36:25 GMT 2017

    Bug fixes:
    - Make perllocal.pod files reproducible
    - META_ADD/MERGE default meta version based on each other
    - Eliminate an ancient, unneeded, dangerous call to Carp::longmess

+++ cpan/ExtUtils-MakeMaker/t/lib/MakeMaker/Test/Setup/XS.pm
@@ -19,7 +19,8 @@ use ExtUtils::MM;
 my $tempdir = tempdir(DIR => getcwd, CLEANUP => 1);
 chdir $tempdir;
 my $typemap = 'type map';
-$typemap =~ s/ //g unless MM->new({NAME=>'name', NORECURS=>1})->can_dep_space;
+my $MM = MM->new({NAME=>'name', NORECURS=>1});
+$typemap =~ s/ //g unless $MM->can_dep_space;
 chdir File::Spec->updir;
 
 my $PM_TEST = <<'END';
@@ -148,6 +149,7 @@ $label2files{static} = +{
     $MAKEFILEPL, 'Test', 'lib/XS/Test.pm', qq{'$typemap'},
     q{LINKTYPE => 'static'},
   ),
+  "blib/arch/auto/share/dist/x-y/libwhatevs$MM->{LIB_EXT}" => 'hi there', # mimic what File::ShareDir can do
 };
 
 $label2files{subdirs} = +{
@@ -210,6 +212,69 @@ $label2files{subdirsstatic} = +{
     q{DEFINE => '-DINVAR=input', LINKTYPE => 'static',},
   ),
 };
+
+# to mimic behaviour of CGI-Deurl-XS version 0.08
+$label2files{subdirsskip} = +{
+  %{ $label2files{subdirscomplex} }, # make copy
+  'Makefile.PL' => sprintf(
+    $MAKEFILEPL,
+    'Test', 'Test.pm', qq{},
+    <<'EOF',
+MYEXTLIB => 'Other$(DIRFILESEP)libparser$(LIB_EXT)',
+EOF
+  ) . <<'EOF',
+sub MY::postamble {
+    my ($self) = @_;
+    return '$(MYEXTLIB) : Other$(DIRFILESEP)Makefile'."\n\t".$self->cd('Other', '$(MAKE) $(PASSTHRU)')."\n";
+}
+EOF
+  'Other/Makefile.PL' => sprintf(
+    $MAKEFILEPL,
+    'Other', 'Other.pm', qq{},
+    <<'EOF',
+SKIP   => [qw(all static dynamic )],
+clean  => {'FILES' => 'libparser$(LIB_EXT)'},
+EOF
+  ) . <<'EOF',
+sub MY::top_targets {
+  my ($self) = @_;
+  my $static_lib_pure_cmd = $self->static_lib_pure_cmd('$(O_FILES)');
+  <<'SNIP' . $static_lib_pure_cmd;
+all :: static
+
+pure_all :: static
+
+static :: libparser$(LIB_EXT)
+
+libparser$(LIB_EXT): $(O_FILES)
+SNIP
+}
+EOF
+  't/plus1.t' => <<'END',
+#!/usr/bin/perl -w
+use Test::More tests => 2;
+use_ok "XS::Test";
+is XS::Test::plus1(3), 4;
+END
+  'Test.xs' => <<EOF,
+#ifdef __cplusplus
+extern "C" {
+#endif
+int plus1(int);
+#ifdef __cplusplus
+}
+#endif
+$XS_TEST
+int
+plus1(input)
+       int     input
+   CODE:
+       RETVAL = plus1(input);
+   OUTPUT:
+       RETVAL
+EOF
+};
+virtual_rename('subdirsskip', 'Other/lib/file.c', 'Other/file.c');
 
 my $XS_MULTI = $XS_OTHER;
 # check compiling from top dir still can include local
@@ -326,7 +391,7 @@ sub list_dynamic {
     $^O ne 'MSWin32' ? (
         [ 'bscode', '', '' ],
         [ 'bscodemulti', '', '' ],
-        $^O ne 'VMS' ? ([ 'subdirscomplex', '', '' ]) : (),
+        $^O !~ m!^(VMS|aix)$! ? ([ 'subdirscomplex', '', '' ]) : (),
     ) : (), # DynaLoader different
     [ 'subdirs', '', '' ],
     [ 'subdirsstatic', ' LINKTYPE=dynamic', ' LINKTYPE=dynamic' ],
@@ -335,6 +400,7 @@ sub list_dynamic {
     [ 'staticmulti', ' LINKTYPE=dynamic', ' LINKTYPE=dynamic' ],
     [ 'staticmulti', ' dynamic', '_dynamic' ],
     [ 'xsbuild', '', '' ],
+    [ 'subdirsskip', '', '' ],
   );
 }
 
@@ -391,7 +457,11 @@ sub run_tests {
   }
 
   chdir File::Spec->updir or die;
-  ok rmtree($dir), "teardown $dir";
+  if ($ENV{EUMM_KEEP_TESTDIRS}) {
+    ok 1, "don't teardown $dir";
+  } else {
+    ok rmtree($dir), "teardown $dir";
+  }
 }
 
 1;
